+++ src/components/Constants/team.js
@@ -1,0 +1,33 @@
+import Devendra from "../../assets/images/Team/devendra.svg"
+import Nitya from "../../assets/images/Team/nitya.svg"
+import Subroto from "../../assets/images/Team/subroto.svg"
+import Founder from "../../assets/images/Team/Founder.svg"
+import Designer from "../../assets/images/Team/Designer.svg"
+
+const teams = [
+    {
+        name: "Devendra Singh Rana",
+        position: "Founder",
+        since: "2023",
+        image: Devendra,
+        icon:Founder
+    },
+    {
+        name: "Nitya Dwivedi",
+        position: "Co-Founder",
+        since: "2023",
+        image: Nitya,
+        icon:Founder
+
+    },
+    {
+        name: "Subroto Santra",
+        position: "UI/UX Designer",
+        since: "2023",
+        image: Subroto,
+        icon:Designer
+
+    }
+];
+
+export default teams;
